Source clean up

+++ submods/fermion-varia/src/main/java/grondag/fermion/client/models/SimpleModels.java
@@ -1,12 +1,12 @@
 /*******************************************************************************
  * Copyright 2019 grondag
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License.  You may obtain a copy
  * of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
@@ -51,8 +51,8 @@ public class SimpleModels implements ModelVariantProvider {
 	 * coordinates.
 	 */
 	public static void contractUVs(int spriteIndex, Sprite sprite, MutableQuadView poly) {
-		final float uPixels = (float) sprite.getWidth() / (sprite.getMaxU() - sprite.getMinU());
-		final float vPixels = (float) sprite.getHeight() / (sprite.getMaxV() - sprite.getMinV());
+		final float uPixels = sprite.getWidth() / (sprite.getMaxU() - sprite.getMinU());
+		final float vPixels = sprite.getHeight() / (sprite.getMaxV() - sprite.getMinV());
 		final float nudge = 4.0f / Math.max(vPixels, uPixels);
 
 		final float u0 = poly.spriteU(0, spriteIndex);
